+++ KSP2Runtime/KSPVessel/KSPVesselModule.ModuleFairing.cs
@@ -30,7 +30,7 @@ public partial class KSPVesselModule {
             get => dataFairing.FairingEnabled.GetValue();
             set => dataFairing.FairingEnabled.SetObject(value);
         }
-        
+
         [KSMethod]
         public bool Jettison() {
             if (!KSPContext.CurrentContext.Game.SpaceSimulation.TryGetViewObject(part.SimulationObject,