+++ CHANGES.md
@@ -22,6 +22,8 @@
   * `heatshield.is_ablating`
   * `heatshield.is_ablator_exhausted`
   * `heatshield.ablator_ratio` (0.0 - 1.0)
+* Add propellant information to engine via `engine.current_propellant` and `engine.propellants`
+  * Extend `ResourceDefinition` by `is_recipe` and `recipe_ingredients`
 
 ## 0.5.3.3 -> 0.5.3.4